microservices/lesson-service/requirements.txt
# Django Lesson Generation Service
# Port: 8003
# Purpose: PDF processing, AI lesson generation, user-specific data management

Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1

# PDF Processing
PyMuPDF==1.23.14
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# AI/LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.8
openai==1.3.7
google-generativeai==0.3.2

# Database
pymongo==4.6.0
mongoengine==0.27.0

# API & Communication
requests==2.31.0
pika==1.3.2

# Utilities
colorama==0.4.6
python-dotenv==1.0.0